backend/pyproject.toml
# Hospital Backend - Python Dependencies
# Managed with uv: https://github.com/astral-sh/uv
# Install: curl -LsSf https://astral.sh/uv/install.sh | sh
# Usage: uv pip install -r requirements.txt

[project]
name = "hospital-backend"
version = "0.1.0"
description = "Enterprise hospital management backend"
authors = [{name = "Immanuel Njogu"}]
requires-python = ">=3.12"
license = {text = "Custom Commercial License"}

dependencies = [
    "Django>=5.0,<5.1",
    "djangorestframework>=3.14.0",
    "django-environ>=0.11.0",
    "psycopg2-binary>=2.9.9",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.27.0",
    "django-filter>=23.5",
    "django-cors-headers>=4.3.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.12.0",
    "django-extensions>=3.2.0",
    "django-debug-toolbar>=4.2.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "django-stubs>=4.2.0",
    "djangorestframework-stubs>=3.14.0",
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
    "hypothesis>=6.92.0",
    "bandit>=1.7.5",
    "safety>=2.3.5",
]

prod = [
    "gunicorn>=21.2.0",
    "uvicorn[standard]>=0.24.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "sentry-sdk>=1.38.0",
    "django-redis>=5.4.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=20.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
known_django = ["django"]
known_drf = ["rest_framework"]
sections = ["FUTURE", "STDLIB", "DJANGO", "DRF", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--reuse-db --cov=apps --cov-report=html --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
